+++ src/ModuleResolver.ts
@@ -1,3 +1,6 @@
+import { execSync } from "child_process";
+import { trace } from "console";
+import * as fs from "fs";
 import * as mem from "mem";
 import * as path from "path";
 import * as prettier from "prettier";
@@ -6,10 +9,11 @@ import * as resolve from "resolve";
 import * as semver from "semver";
 // tslint:disable-next-line: no-implicit-dependencies
 import { Disposable } from "vscode";
+import { resolveGlobalNodePath, resolveGlobalYarnPath } from "./Files";
 import { LoggingService } from "./LoggingService";
 import { FAILED_TO_LOAD_MODULE_MESSAGE } from "./message";
 import { NotificationService } from "./NotificationService";
-import { PrettierModule } from "./types";
+import { PackageManagers, PrettierModule } from "./types";
 import { getConfig, getWorkspaceRelativePath } from "./util";
 
 const minPrettierVersion = "1.13.0";
@@ -19,6 +23,43 @@ declare const __non_webpack_require__: typeof require;
 interface ModuleResult {
   moduleInstance: any | undefined;
   modulePath: string | undefined;
+}
+
+const globalPaths: {
+  [key: string]: { cache: string | undefined; get(): string | undefined };
+} = {
+  npm: {
+    cache: undefined,
+    get(): string | undefined {
+      return resolveGlobalNodePath(trace);
+    }
+  },
+  pnpm: {
+    cache: undefined,
+    get(): string {
+      const pnpmPath = execSync("pnpm root -g")
+        .toString()
+        .trim();
+      return pnpmPath;
+    }
+  },
+  yarn: {
+    cache: undefined,
+    get(): string | undefined {
+      return resolveGlobalYarnPath(trace);
+    }
+  }
+};
+
+function globalPathGet(packageManager: PackageManagers): string | undefined {
+  const pm = globalPaths[packageManager];
+  if (pm) {
+    if (pm.cache === undefined) {
+      pm.cache = pm.get();
+    }
+    return pm.cache;
+  }
+  return undefined;
 }
 
 export class ModuleResolver implements Disposable {
@@ -46,7 +87,7 @@ export class ModuleResolver implements Disposable {
       return prettier;
     }
 
-    const { prettierPath } = getConfig();
+    const { prettierPath, packageManager } = getConfig();
 
     // tslint:disable-next-line: prefer-const
     let { moduleInstance, modulePath } = this.requireLocalPkg<PrettierModule>(
@@ -54,6 +95,17 @@ export class ModuleResolver implements Disposable {
       "prettier",
       prettierPath
     );
+
+    if (!moduleInstance) {
+      const globalModuleResult = this.requireGlobalPkg<PrettierModule>(
+        packageManager,
+        "prettier"
+      );
+      if (globalModuleResult.moduleInstance && globalModuleResult.modulePath) {
+        moduleInstance = globalModuleResult.moduleInstance;
+        modulePath = globalModuleResult.modulePath;
+      }
+    }
 
     if (!moduleInstance && isUserInteractive) {
       this.loggingService.logInfo("Using bundled version of prettier.");
@@ -84,8 +136,21 @@ export class ModuleResolver implements Disposable {
     return moduleInstance || prettier;
   }
 
-  public getModuleInstance(fsPath: string, pkgName: string): any {
-    const { moduleInstance } = this.requireLocalPkg<any>(fsPath, pkgName);
+  public getModuleInstance(
+    fsPath: string,
+    packageManager: PackageManagers,
+    pkgName: string
+  ): any {
+    let { moduleInstance } = this.requireLocalPkg<any>(fsPath, pkgName);
+    if (!moduleInstance) {
+      const globalModuleResult = this.requireGlobalPkg<PrettierModule>(
+        packageManager,
+        "prettier"
+      );
+      if (globalModuleResult.moduleInstance) {
+        moduleInstance = globalModuleResult.moduleInstance;
+      }
+    }
     return moduleInstance;
   }
 
@@ -146,6 +211,35 @@ export class ModuleResolver implements Disposable {
       ]);
     }
     return { moduleInstance: undefined, modulePath };
+  }
+
+  private requireGlobalPkg<T>(
+    packageManager: PackageManagers,
+    pkgName: string
+  ): ModuleResult {
+    const resolvedGlobalPackageManagerPath = globalPathGet(packageManager);
+    if (resolvedGlobalPackageManagerPath) {
+      const modulePath = path.join(resolvedGlobalPackageManagerPath, pkgName);
+      if (fs.existsSync(modulePath)) {
+        try {
+          const moduleInstance = this.loadNodeModule(modulePath);
+          if (this.resolvedModules.indexOf(modulePath) === -1) {
+            this.resolvedModules.push(modulePath);
+          }
+          this.loggingService.logInfo(
+            `Loaded module '${pkgName}@${moduleInstance.version}' from '${modulePath}'`
+          );
+          return { moduleInstance, modulePath };
+        } catch (error) {
+          this.loggingService.logError(
+            `Failed to load ${pkgName} from '${modulePath}'`,
+            error
+          );
+          return { moduleInstance: undefined, modulePath };
+        }
+      }
+    }
+    return { moduleInstance: undefined, modulePath: undefined };
   }
 
   // Source: https://github.com/microsoft/vscode-eslint/blob/master/server/src/eslintServer.ts#L209