+++ src/types.d.ts
@@ -4,6 +4,8 @@ type PrettierModule = typeof prettier;
 
 type TrailingCommaOption = "none" | "es5" | "all";
 
+export type PackageManagers = "npm" | "yarn" | "pnpm";
+
 /**
  * prettier-vscode specific configuration
  */
@@ -40,10 +42,13 @@ interface IExtensionConfig {
    */
   requireConfig: boolean;
   /**
+   * The package manager to use when resolving global modules.
+   */
+  packageManager: PackageManagers;
+  /**
    * Array of language IDs to ignore
    */
   disableLanguages: string[];
-
   /**
    * If true, take into account the .editorconfig file when resolving configuration
    */